chore: rename Pickedup to PickedUp
